requirements.txt
antlr4-python3-runtime==4.13.1
matplotlib==3.8.2
flask==2.3.3
werkzeug==2.3.7
itsdangerous==2.1.2
jinja2==3.1.2
click==8.1.7
markupsafe==2.1.3
numpy==1.26.2
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
cycler==0.12.1
kiwisolver==1.4.5
fonttools==4.44.0
packaging==23.2 